refactor(fe): modularize landing worker
