+++ frontend/landing/public/worker/response.js
@@ -1,0 +1,8 @@
+export const json = (body, init = {}) =>
+  Response.json(body, {
+    ...init,
+    headers: {
+      'Cache-Control': 'no-store',
+      ...(init.headers || {}),
+    },
+  })
